Docker-compose with a PostgreSQL database and Adminer interface.
A command like:
  $ docker-compose up
from the repository's root directory will build and start the Docker
containers for a Postgres database and an Adminer web interface.
Postgres is accessible through the command line client on port 2345
  $ psql -h localhost -p 2345 -d docker -U docker --password
with a password of 'docker'.
Adminer is accessible through a browser pointed at localhost:80.

.env file is automatically sourced by docker-compose to pick up
environment variables. To check the correctness of the configuration
  $ docker-compose config
will show the exact configuration used by docker-compose.