Add a MongoDB database and Adminer interface to docker-compose.
Mongo: Use a default mongo image from Docker Hub.
       Use 'docker' for username and password.
       Default Mongo port of 27017 is used (command line access).

Adminer: Mongo needs a custom image that includes PHP5.
       The Adminer image used for Postgres is PHP7, so it will
       not work with MongoDB. That is why there are now two
       Adminer services.

Postgres Adminer web access is on http://localhost:8000
MongoDB Adminer web access is on http://localhost:8080

The .env file in the root directory of the repository has all the
entries needed for logging in to each database.
The Postgres Adminer automatically fills in these values, but the
MongoDB Adminer does not, and will need to be typed in by the user.