Hook up a basic connection from Python3 to the Postgres database.
Docker image is build in the transaction-server subdirectory.
apt-get and pip3 install some dependencies for the Python
library psycopg2.

The database.ini is parsed by the configparser library, and is a
critical part of making this all work.

The most difficult part of this commit was figuring out that
Docker makes its own network, with each node named for its
service name from docker-compose.yml. So the "host" in database.ini
is called "postgres", not "localhost" as you would expect, so that
it can make a connection to the postgres docker container to access
the database. Very tricky.