Frontend: Query the Transaction Server for movie data.
Previously the frontend talked directly to the MongoDB database.
Now the frontend requests are sent to the transaction server
which will do the talking to the databases. Frontend doesn't know.

Movie (and actor) search is not addressed in this commit. More
work in the transaction server API is needed for search to work.

A url variable is used for the transaction server movies endpoint.

The "requests" package is used for HTTP GET/POST requests.