Trans-srv to sql: Initialize SQL tables via init.sql file.
The Postgres Docker image has builtin functionality to initialize
the database via a supplied init.sql file. This file must be placed
in /docker-entrypoint-initdb.d/ directory inside the Docker container.

A Dockerfile is now used in the postgres directory to copy over this
file as well as set up some default options for Postgres.

The SQL schema is currently only a 'movies' table.
Postgres will auto-increment a generated serial number id.
movie_name and movie_date must not be NULL, and must be unique.
movie_desc is optional, and consists of unlimited freeform text.