Adding a movie via endpoint adds it to Postgres.
Both workload generator and frontend now add a movie record to
the Postgres database, instead of the builtin Python dictionary.

The database insert can be seen via Adminer on port 8000.

Adding a movie via the frontend works completely, due to some care
taken to make sure movie_desc is null (or None) if it is not supplied.
The frontend date field for adding a movie now enforces that a number
is supplied (as apposed to being freeform text). Postgres is happy,
since the database schema specifies that it must be an integer.

The return value from the database insert still needs to be checked
for success, so that the record can be written to the Mongo database.