Transaction Server: Hook in MongoDB (via async Motor library).
This commit has many debug messages, but its purpose is to be a
reference for an async connection to the MongoDB database.

The crucial part of getting the async Motor library to work is for
it to connect into the event loop that Quart creates and manages.
This is done via the @app.before_serving annotation on an init()
method, and a call to asyncio.get_event_loop() while in init().
If get_event_loop() is called anywhere else, it will create its
own event loop, which does not work the way we expect (can't have
two event loops in the application).

Any call on the Motor client or collection must be awaited, such as
insert_one(), find_one(), and list_database_names().

Add the mongo database service as a dependency of the trans-srv
transaction server service in the docker-compose.yml file.