Transaction Server: Flesh out all Mongo database writes and reads.
A GET request to endpoint /api/movies will return, in JSON,
all records from the Mongo database, (currently) up to 100 records.
A GET request to endpoint /api/movies/[movie_id] will return, in JSON,
the record in the Mongo database that has an _id of movie_id.

The serialized id returned from Postgres is used as the special "_id"
in Mongo, which guarantees that there will not be any duplicate entries.

Many debug print() statements are left in for now.

The call to "cursor = coll.find()" does not need an await, as it only
connects to the Mongo database when the cursor is used, such as in
"movie_list = await cursor.to_list()".

Inserting into Mongo will only happen if the insert to Postgres is
successful, by use of the except catch block, which will skip the
call to insert_to_mongo().

The sample "movies" dictionary database used for testing is no longer needed.

The get_movies() method that reads from Postgres is not used,
and is renamed to get_movies_from_postgres(). Keep it for diagnostics.