Transaction Server: Add search URL query parameters to GET API.
A GET request to the endpoint /api/movies can now have optional
query parameters for name, date, and desc.
An example query is as follows:
  http://localhost:5555/api/movies?name=Unchained&date=2009&desc=Cool
Each parameter is optional, and they further restrict (filter) the
returned database records.
The searching is based on an exact, case sensitive, match.
A search like
  http://localhost:5555/api/movies?date=2000
will return all movies that are from the year 2000.

Frontend changes for search functionality will be in another commit.