Modified wl_generator.cpp: - Added factory function.  It creates lambda functions that make http get requests to the transaction server, which is the method used to query the database for records. - Added function that creates a series of the above-mentioned lambda functions and schedules them with a thread pool.
TODO: timing is problem atm.  Since the read and write requests are currently being made by the same thread pool, the ordering of each requests is indeterminate.  The main function must be restructured in such a way that it can be guaranteed that certain tasks are completed before others are started.  Ie, if records #1 through record #10 are being added, then queries are made searching for records #1 through record #10, it is critical that all writes are complete before the reads occur.
This type of atomicity can be achieved by segregating certain batches of tasks into different thread pools, and ensuring the destruction of one thread pool before creating the next.
ie - the destruction of a thread pool causes every thread within to be joined by the main thread of exeuction.  This in turn, insures that all tasks in the thread pool are complete before the main thread of execution proceeds.