Frontend: Add search by using transaction server search API.
Previously the frontend contacted the Mongo database directly to
do a search via the search box on the homepage.
Now, the frontend communicates only with the transaction server,
and uses the newly minted search features of the API.

The semantics of frontend search is an empty key in the "Search for ..."
box will request all of the records from the transaction server.
A search using the drop-down of either movie name, date, or
description and a key entered into the "Search for ..." box will
send a GET request to the transaction server with the parameters
filled in for the URL query string.

Don't import bson and pymongo; remove them from requirements.txt.